리뷰 더 불러오기 api 추가

+++ apis/review/reviewResolver.py
@@ -15,14 +15,14 @@ class Create_review(Resource):
         # try:
         # logined_user = User.query.filter_by(email=session['login']).first()
         # user_id = logined_user.id
-        # if not session:
-        #     abort(500, "로그인 해주세요")
-        # else:
-        user_id = request.json['user_id']
-        liquor_id = request.json['liquor_id']
-        rating = request.json['rating']
-        content = request.json['content']
-        return reviewService.create_review(user_id,liquor_id,rating,content)
+        if not session:
+            abort(500, "로그인 해주세요")
+        else:
+            user_id = request.json['user_id']
+            liquor_id = request.json['liquor_id']
+            rating = request.json['rating']
+            content = request.json['content']
+            return reviewService.create_review(user_id,liquor_id,rating,content)
         # except:
         #     abort(500, "리뷰 등록 실패.")
 
@@ -33,24 +33,24 @@ class Update_review(Resource):
     @Review.response(500, 'Failed to revise the review')
     def post(self):
         '''술 리뷰 수정'''
-        # if not session:
-        #     abort(500, "로그인 해주세요")
-        # else:
-        review_id = request.json['review_id']
-        user_id = request.json['user_id']
-        rating = request.json['rating']
-        content = request.json['content']
-        return reviewService.update_review(user_id,review_id,rating,content)
+        if not session:
+            abort(500, "로그인 해주세요")
+        else:
+            review_id = request.json['review_id']
+            user_id = request.json['user_id']
+            rating = request.json['rating']
+            content = request.json['content']
+            return reviewService.update_review(user_id,review_id,rating,content)
         # except:
         #     abort(500, "로그인 해주세요.")
 
 @Review.route("/delete/<int:review_id>/user/<int:user_id>")
 class delete_review(Resource):
     def delete(self, user_id, review_id):
-        # if not session:
-        #     abort(500, "로그인 해주세요")
-        # else:
-        return reviewService.delete_review(user_id, review_id)
+        if not session:
+            abort(500, "로그인 해주세요")
+        else:
+            return reviewService.delete_review( user_id, review_id)
 
 
 @Review.route("/<int:review_id>/user/<int:user_id>")
@@ -61,4 +61,14 @@ class review(Resource):
     @Review.marshal_with(show_reviewDTO, mask=False)
     def get(self, user_id, review_id):
         '''리뷰 하나만 조회'''
-        return reviewService.show_review(user_id, review_id)
+        return reviewService.get_review(user_id, review_id)
+
+@Review.route("/more_reviews/liquor/<int:liquor_id>/<int:last_review_id>")
+class next_reviews(Resource):
+    @Review.response(200, "Review exist",next_reviewsDTO)
+    @Review.response(404, "Not found")
+    @Review.response(500, "Review doesn't exist")
+    @Review.marshal_with(next_reviewsDTO, mask=False)
+    def get(self, liquor_id, last_review_id):
+        '''다음 리뷰 불러오기'''
+        return reviewService.get_next_reviews(liquor_id, last_review_id)
